Refactor `xla_dist.py`: breakout `worker.py` and `cluster.py`
